+++ pczt/src/roles.rs
@@ -10,7 +10,7 @@
 
 pub mod creator;
 
-#[cfg(any(feature = "io-finalizer", feature = "signer"))]
+#[cfg(any(feature = "io-finalizer", feature = "signer", feature = "tx-extractor"))]
 pub(crate) mod tx_data;
 
 #[cfg(feature = "io-finalizer")]